+++ demos/polish_docker_nodes.py
@@ -1,0 +1,8 @@
+from plumbery.engine import PlumberyEngine
+
+PlumberyEngine('fittings.yaml').polish_nodes('docker', 'rub')
+
+
+
+
+