+++ demos/polish_sql_nodes.py
@@ -1,6 +1,6 @@
 from plumbery.engine import PlumberyEngine
 
-PlumberyEngine('fittings.yaml').polish_nodes('sql', 'ansible')
+PlumberyEngine('fittings.yaml').polish_nodes('sql', 'spit')