+++ demos/start_docker_nodes.py
@@ -1,0 +1,8 @@
+from plumbery.engine import PlumberyEngine
+
+PlumberyEngine('fittings.yaml').start_nodes('docker')
+
+
+
+
+