+++ plumbery/polishers/rub.update.sh
@@ -1,0 +1,4 @@
+#!/usr/bin/env bash
+
+# ensure we run most recent versions of software
+apt-get -y update && apt-get -y upgrade