+++ HISTORY.rst
@@ -3,6 +3,20 @@
 History
 -------
 
+0.2.0 (2015-11-29)
+---------------------
+
+* Allow for node bootstrapping via SSH
+* Push SSH public key
+* Update Linux packages
+* Install Docker
+* Provide an inventory of running nodes
+* Build inventory for ansible
+* Extension mechanism called polishers
+* Add demonstration scripts related to polishers
+* Test coverage 59%
+* fake8 complains only about long lines and line termination
+
 0.1.2 (2015-11-27)
 ---------------------